requirements.txt
fastapi
uvicorn[standard]
chainlit
httpx
langchain
langchain-groq
tavily
python-dotenv
psycopg2[binary]
langsmith